requirements.txt
Flask
tensorflow
Pillow
gunicorn
Werkzeug
flask-cors